chore: restructure local components
